# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Onereach API",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "abiodun.adeyinka@telvida.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/apponereach/public",
            "description": "Onereach API Server"
        }
    ],
    "paths": {
...
        "/api/app/auth/password": {
            "put": {
                "tags": [
                    "Auth"
                ],
                "summary": "Change Password",
                "description": "Change Password",
                "operationId": "changePassword",
                "parameters": [
                    {
                        "name": "current_password",
                        "in": "path",
                        "description": "Current User Password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "path",
                        "description": "New User Password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password_confirmation",
                        "in": "path",
                        "description": "New User Password Confirmation",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "User password is has been changed"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
        "/api/app/auth/password/reset": {
            "put": {
                "tags": [
                    "Auth"
                ],
                "summary": "Reset Password",
                "description": "Reset Password",
                "operationId": "resetPassword",
                "parameters": [
                    {
                        "name": "email",
                        "in": "path",
                        "description": "User email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "account",
                        "in": "path",
                        "description": "Company short or account name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "User password is has been changed"
                    },
                    "200": {
                        "description": "No user exists for this phonenumber"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
        "/api/app/billing/order/cart": {
            "get": {
                "tags": [
                    "Billing Order"
                ],
                "summary": "Get cart information",
                "description": "Return data in the cart",
                "operationId": "getCart",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/app/billing/subscription/current": {
            "get": {
                "tags": [
                    "Billing Subscription"
                ],
                "summary": "Get Current Subscription ",
                "description": "Current Subscription Details",
                "operationId": "App\\Http\\Controllers\\Api\\Tenant\\Billing\\SubscriptionController::currentSubscription",
                "responses": {
                    "200": {
                        "description": "Successful"
                    },
                    "401": {
                        "description": "Unautorized"
                    }
                }
            }
        },
        "/api/app/billing/subscription/current/procesing": {
            "get": {
                "tags": [
                    "Billing Subscription"
                ],
                "summary": "Get Current Incomplete Subscription ",
                "description": "Processing Subscription ",
                "operationId": "App\\Http\\Controllers\\Api\\Tenant\\Billing\\SubscriptionController::currentProcessingSubscription",
                "responses": {
                    "200": {
                        "description": "successfully "
                    },
                    "401": {
                        "description": "Unautorized"
                    }
                }
            }
        },
        "/api/app/registion/save-order": {
            "post": {
                "tags": [
                    "Boarding"
                ],
                "summary": "Save customer order",
                "description": "Returns generated Order and Transaction",
                "operationId": "saveOrder",
                "parameters": [
                    {
                        "name": "subscription_id",
                        "in": "path",
                        "description": "Subscription id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
...
        "/api/app/voucher/claim": {
            "post": {
                "tags": [
                    "Voucher"
                ],
                "summary": "User Can Claim Voucher To Get Discount ",
                "description": "Voucher User",
                "operationId": "App\\Http\\Controllers\\Api\\Tenant\\VoucherController::claimVoucher",
                "parameters": [
                    {
                        "name": "voucher_code",
                        "in": "path",
                        "description": "Voucher code with 12 characters",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "discount successfully applied"
                    },
                    "442": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
        "/api/app/voucher/info": {
            "get": {
                "tags": [
                    "Voucher"
                ],
                "summary": "Get Voucher Information to see applicable discount. ",
                "description": "Voucher User",
                "operationId": "App\\Http\\Controllers\\Api\\Tenant\\VoucherController::getVoucherData",
                "parameters": [
                    {
                        "name": "voucher_code",
                        "in": "path",
                        "description": "Voucher code with 12 characters",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Summary infomation about the voucher and amount of discount attached."
                    },
                    "442": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "Password Based": {
                "type": "oauth2",
                "description": "Use a global client_id / client_secret and your username / password combo to obtain a token",
                "in": "header",
                "flows": {
                    "password": {
                        "authorizationUrl": "/oauth/authorize",
                        "tokenUrl": "/oauth/token",
                        "refreshUrl": "/oauth/token/refresh",
                        "scopes": []
                    }
                },
                "scheme": "https"
            }
        }
    },
...
}